// src/components/reception/graphql/types.ts
export type Service = {
  __typename: string
  id: string | null
  name?: string | null
  price?: number | null
  bhytId?: string | null
  bhytPrice?: number | null
  bhytName?: string | null
  chooseSpecIndex?: boolean | null
  clinicId?: string | null
  cost?: number | null
  createAt?: string | null
  createBy?: string | null
  describe?: string | null
  gender?: number | null
  insurancePaymentRate?: number | null
  serviceGroupId?: string | null
  serviceTypeId?: string | null
  status?: boolean | null
  quantity?: number | null
  departmentId?: string | null
}

export type ServiceGroup = {
  name?: string
  serviceGroupId: string
  services: Service[]
}